ai/requirements.txt
google-generativeai>=0.3.0
typing-extensions>=4.0.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=0.21.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
faiss-cpu>=1.7.4
langchain-community>=0.0.25
waitress
python-docx>=0.8.11
Pillow>=10.0.1
pytesseract>=0.3.10